Feature/refactor rating prediction (#1920)
* team analyzer filter handles challenger cup as friendlies now

* logging RatingPredictionManager

* Speciality.java->specialty

* rating logging

* refactoring RatingPredictionManager

* #1782 start schum rating

* #1782 schum rating

* #1782 RatingPredictionManager teamspirit formula

* #1805 release_notes.md

* typo

* #1782 rating refactoring

* #1782 rating refactoring

* #1782 rating refactoring

* #1782 add experience calculations to strenght cache

* init ratingprediction.json

* json preciction definition

* ratingprediction.json

* start HO7.1

* start HO8.0

* continue

* Release/7 (#1859)

* start HO7.1

* #1856 fix currency of secondary teams

* #1856 fix currency of secondary teams

---------


* #1857 fix playername getter in MatchHighlightsTable (#1860)


* #1861 fix exception in GoalDiffCriteria (#1864)


* Fix click on “Last Match” column when it has been moved. (#1863)

The “Last Match” column is clickable, and opens:
- The match in Hattrick when shift-clicking,
- The match in HO when double-clicking.

This commit also adds some comment to clarify the behaviour.

* Minor UI improvement in teamAnalyzer. (#1867)

This adds a bit of padding around the check boxes on the teamAnalyser panel,
and centers the automatic / manual radio buttons.

It also cleans up code a bit, in particular in lineup assistant.  Probably more
effort to come on that front.

* #1858 (#1869)

* #1858 debug date time parsing

* #1858 debug currency parsing

* #1858 using Helper.formatCurrency

* #1858 Player.baseWage reset setter getter (not used for now, but senseful for the future)

* #1858 Player.baseWage reset setter getter (not used for now, but senseful for the future)

* #1858 translate comments in Helper.java

* #1858 remove  unused imports in HODateTime.java (add some comments)

* #1858 fix Player.setExpiryDate, use localized DateTimeFormatter

* #1858 fix PlayerConverter add comment explaining offsSpecialty

* #1858 fix PlayerConverter remove old commented methods

* #1858 fix PlayerConverter.parseLocalDateTime

* #1858 PlayerConverter scan wage

* #1858 PlayerConverter scan skill refactored

* #1858 PlayerConverter scan wage
fix indexRowWarning

* Fixes deadline date parsing in Transfer Scout.

---------


* refactor rating prediction

* #1873 (#1874)

* #1873 fix MatchAndLineupSelectionPanel.updateStyleOfPlayComboBox nt team selection

* #1873 release_notes.md

* ratingprediction formulas (wip)

* #1878 (#1879)

* fix OnlineWorker.downloadLastLineup

* MyConnector.downloadPlayerDetails include match info

* #1878 release_notes.md

* #1878 show last match date of nt team players

* RatingPredictionManager remove debug logs

* new class RatingPredictionModel

* new class RatingPredictionModel

* #1881 (#1882)

* some cleanup

* #1881 add undocumented goal difference criteria

* #1881 add undocumented goal difference criteria release_notes.md

* #1881 debug worlflow (#1883)

* some cleanup

* #1881 add undocumented goal difference criteria

* #1881 add undocumented goal difference criteria release_notes.md

* #1881 fix workflow

* RatingPredictionModel (wip)

* Release ho7.1 (#1887)

* release_notes.md update translaation statistics

* release_notes.md

* release HO 7.1

* release_notes.md

* RatingPredictionModel

* #1889 download new trainer info from player details (#1890)

* RatingPredictionModel (wip)

* RatingPredictionModel (wip)

* stable 7.2 (#1893)

* rating prediction (wip)

* rating prediction (wip)

* #1899 (#1900)

* #1899 training recalculation recovers deleted skill ups

* #1901 PlayerDetailsPanel reset homegrown icon (#1902)

* RatingPredictionModel (wip)

* RatingPredictionModel (wip)

* RatingPredictionModel (wip)

* RatingPredictionModel (wip)

* RatingPredictionModel (wip)

* RatingPredictionModel (wip)

* calcTeamSpirit

* fix RatingChartData

* fix RatingChartData

* fix RatingChartData

* #1849 (#1912)

* ModuleConfigTable.saveConfig

* Update install4j version. (#1906)

This is an attempt at fixing issue #1849 by updating install4j to the latest
version, as it is unclear at this point what is causing the install to hang on
some MacOS machines.

(cherry picked from commit d2a5b0b2c8d2a08d2aa71082ef848ca88b021035)

* release-dev INSTALL4J_HOME (#1907)

(cherry picked from commit 628d6856a60b5ca3c0438246b82801b0e91e6e70)

* https://github.com/ho-dev/HattrickOrganizer (#1911)

(cherry picked from commit 823bfa0c8e5649503bbd0c55f019215078e4bd2c)

* #1849 release_notes.md

---------


* wip

* #1913 fix Player.getMotherclubId

* fit values of RatingPredictionModel.calcSector

* RatingPredictionModel.calcRatingSectorScale

* RatingPredictionModel.calcPlayerScale (wip)

* Player.getPositionRating

* copyrights

* getPlayerMatchAverageRating

* fix edit SubstitutionEditView hatstats update

* fix Substitution.pos

* fix Substitution.setRoleId

* fix RatingPredictionModel.calcStamina

* tune RatingPredictionModel.getRatingSectorScaleFactor

* fix enum coding style

* read.me

* adjust player rating to hatstats formula (factor 3 for midfield contribution)

* implement schum rank

* tychobrailleur's review

* release_notes.md

---------

+++ src/main/java/core/training/TrainingPerPlayer.java
@@ -34,9 +34,9 @@ public class TrainingPerPlayer  {
     }
 
     /**
-     * Getter for property spieler.
+     * Getter for property player.
      *
-     * @return Value of property spieler.
+     * @return Value of property player.
      */
     public final Player getPlayer() {
         return this._Player;
@@ -77,7 +77,7 @@ public class TrainingPerPlayer  {
 		return experienceSub;
 	}
 
-    public float calcSubskillIncrement(int skill, float skillValueBeforeTraining, HODateTime date) {
+    public float calcSubskillIncrement(int skill, double skillValueBeforeTraining, HODateTime date) {
 
 		int skillValue = (int) skillValueBeforeTraining;
 		float ret = 0;
